+++ js/main.js
@@ -4,4 +4,4 @@ import {SIMILAR_PHOTOS_COUNT} from './constants.js';
 
 const pictures = similarPhotos(SIMILAR_PHOTOS_COUNT);
 randomPicture(pictures);
-import './modalEditingForm.js';
+import './modal-editing-form.js';
